fix and fix.

+++ server/application/controllers/book.php
@@ -4,21 +4,17 @@ class Book extends CI_Controller{
     {
         parent::__construct();
         $this->load->model('mbook/book_service', 'book_service');
-        $this->load->model('muser/session_service', 'book_session_service');
-    }
-    private function session_check()
-    {
-        $data['json'] = $this->book_session_service->islogin();
-        if( $data['json']['code'] != 0 )
-        {
-           $this->load->view('json_view', $data);
-           return;
-        }
+        $this->load->model('muser/login_service', 'login_service');
     }
     public function index()
     {
-        $this->session_check();
-            
+        $data['json'] = $this->login_service->islogin();
+        if( $data['json']['code'] != 0 )
+        {
+            $this->load->view('json_view', $data);
+            return;
+        }
+
         $data['json'] = $this->book_service->get_all_book();
         $this->load->view('json_view', $data);
     }
@@ -26,14 +22,24 @@ class Book extends CI_Controller{
 
     public function view($id)
     {
-        $this->session_check();
+        $data['json'] = $this->login_service->islogin();
+        if( $data['json']['code'] != 0 )
+        {
+            $this->load->view('json_view', $data);
+            return;
+        }
 
         $data['json'] = $this->book_service->get_book($id);
         $this->load->view('json_view', $data);
     }
     public function add()
     {
-        $this->session_check();
+        $data['json'] = $this->login_service->islogin();
+        if( $data['json']['code'] != 0 )
+        {
+            $this->load->view('json_view', $data);
+            return;
+        }
 
         $name = $this->input->post('name');
         $cate = $this->input->post('cate');
@@ -43,10 +49,10 @@ class Book extends CI_Controller{
         if( ! is_numeric($page))
         {
             $data['json'] = array(
-                'code'=>1,
-                'msg'=>'page could not be character, it must be number.',
-                'data'=>''
-            );
+                    'code'=>1,
+                    'msg'=>'page could not be character, it must be number.',
+                    'data'=>''
+                    );
             $this->load->view('json_view', $data);
             return;
         }
@@ -55,14 +61,24 @@ class Book extends CI_Controller{
     }
     public function del($id)
     {
-        $this->session_check();
+        $data['json'] = $this->login_service->islogin();
+        if( $data['json']['code'] != 0 )
+        {
+            $this->load->view('json_view', $data);
+            return;
+        }
 
         $data['json'] = $this->book_service->delete_book($id);
         $this->load->view('json_view', $data);
     }
     public function update($id)
     {
-        $this->session_check();
+        $data['json'] = $this->login_service->islogin();
+        if( $data['json']['code'] != 0 )
+        {
+            $this->load->view('json_view', $data);
+            return;
+        }
 
         $name = $this->input->post('name');
         $cate = $this->input->post('cate');
